refactored array to pull flag names - removed script
greatly simplified the flag generation to remove need for chance.js
(which was generating the country code)

country code is now generated by generating random number from 1 to
197, and going to that position in array to pull the country code,
which is then used as the file name to pull from the folders containing
the flags, and display. Cut out about 20 lines of code, and got rid of
a heavy jquery library.